KMS-27368 separate classes
